Move process_fixcost to E0 instead of E1 to avoid cicular import
