+++ src/multi_vector_simulator/E0_evaluation.py
@@ -176,7 +176,7 @@ def evaluate_dict(dict_values, results_main, results_meta):
             store_result_matrix(dict_values[KPI], dict_values[group][asset])
 
     # Add fix project costs
-    E1.process_fixcost(dict_values)
+    process_fixcost(dict_values)
 
     logging.info("Evaluating key performance indicators of the system")
     E3.all_totals(dict_values)
@@ -293,3 +293,48 @@ def initalize_kpi(dict_values):
             }
         }
     )
+
+
+def process_fixcost(dict_values):
+    r"""
+    Adds fix costs of the project to the economic evaluation of the energy system.
+
+    Parameters
+    ----------
+    dict_values: dict
+        All simulation data with inputs and results of the assets
+
+    Returns
+    -------
+    Updated dict_values with costs attributed in dict values also appended to the dict_values[KPI] (scalar results)
+
+    Notes
+    -----
+
+    Function is tested with:
+    - test_E0_evaluation.test_process_fixcost()
+    """
+    for asset in dict_values[FIX_COST]:
+        # Add parameters that are needed for E2.get_costs()
+        dict_values[FIX_COST][asset].update(
+            {
+                OPTIMIZED_ADD_CAP: {VALUE: 1},
+                INSTALLED_CAP: {VALUE: 0},
+                LIFETIME_PRICE_DISPATCH: {VALUE: 0},
+                FLOW: pd.Series([0, 0]),
+            }
+        )
+
+        E2.get_costs(dict_values[FIX_COST][asset], dict_values[ECONOMIC_DATA])
+        # Remove all parameters that were added before and the KPI that do not apply
+        for key in [
+            OPTIMIZED_ADD_CAP,
+            LIFETIME_PRICE_DISPATCH,
+            INSTALLED_CAP,
+            FLOW,
+            COST_DISPATCH,
+        ]:
+            dict_values[FIX_COST][asset].pop(key)
+        store_result_matrix(
+            dict_values[KPI], dict_values[FIX_COST][asset], fix_cost=True
+        )
